Moving All Function to respected files Hoping for commit review on these commits
