+++ .gitignore
@@ -11,3 +11,4 @@ Initial
 Lua\NES\Fitness.csv
 *auto.state
 *.pool
+*Super Mario Bros. 3*
